feat: 정적 데이터를 constant로 분리합니다.
